+++ includes/IO/AudioFile/impl/lamemp3_Write.h
